Updated Test for Shortest_dis
